server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy>=1.26.0
torch>=2.0.0
rdkit-pypi>=2023.9.1
pydantic>=2.0.0
psycopg2-binary>=2.9.0
transformers>=4.30.0
peft>=0.7.0
chemprop>=2.0.0
